Change Security to interface to provide correct support for injection.
