+++ microprofile/openapi/src/main/java/module-info.java
@@ -30,6 +30,7 @@ module io.helidon.microprofile.openapi {
     requires io.helidon.microprofile.server;
     requires io.helidon.openapi;
     requires jakarta.interceptor.api;
+    requires transitive microprofile.openapi.api;
 
     requires org.jboss.jandex;
