+++ microprofile/server/src/main/java/module-info.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018, 2021 Oracle and/or its affiliates.
+ * Copyright (c) 2018, 2022 Oracle and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -31,7 +31,7 @@ module io.helidon.microprofile.server {
     requires transitive jakarta.cdi;
     requires transitive jakarta.ws.rs;
     requires jakarta.interceptor.api;
-    requires jakarta.json;
+    requires transitive jakarta.json;
     requires io.helidon.jersey.media.jsonp;
 
     requires java.logging;
